Fix remaining Xcode-in-chroot bugs
As of this commit, both Make-based and xcodeproj-based projects
build successfully in the chroot. Hooray!

I also took the opportunity to remove the usage of
DYLD_FORCE_FLAT_NAMESPACE from darwinbuild. As of OS X Lion this
is no longer required, and darwintrace has long since been updated
to use the newer method of interposing methods.